refactor: update button color variants
