+++ src/popup/components/Modals/Confirm.vue
@@ -10,7 +10,7 @@
     </template>
     <template #footer>
       <BtnMain
-        variant="secondary"
+        variant="muted"
         @click="cancel"
       >
         {{ $t('modals.cancel') }}
