+++ src/popup/components/Modals/TransferSend.vue
@@ -22,7 +22,7 @@
     <template #footer>
       <BtnMain
         v-if="showEditButton"
-        variant="secondary"
+        variant="muted"
         text="Edit"
         class="button-action-secondary"
         @click="editTransfer"
@@ -43,8 +43,7 @@
 
 <script>
 import { mapState, mapGetters } from 'vuex';
-import { MODAL_TRANSFER_SEND } from '../../utils/constants';
-import { validateTipUrl } from '../../utils/helper';
+import { MODAL_TRANSFER_SEND, validateTipUrl } from '../../utils';
 import Modal from '../Modal.vue';
 import BtnMain from '../buttons/BtnMain.vue';
 import TransferSendForm from '../TransferSendForm.vue';
