+++ src/styles/variables.scss
@@ -21,6 +21,7 @@ $color-white: #fff;
 $color-black: #000;
 $color-primary: #1161fe;
 $color-primary-hover: #0e52d8;
+$color-secondary: #ff2b5e;
 $color-grey-light: #babac0;
 $color-grey-medium: #373737;
 $color-grey-dark: #787878;
@@ -44,7 +45,7 @@ $color-success: #00ff9d;
 $color-success-dark: #00d3a1;
 $color-success-hover: #00dc87;
 $color-warning: #ffb422;
-$color-danger: #ff2b5e;
+$color-danger: #ff4746;
 
 $color-bg-app: #141414;
 $color-bg-modal: $color-bg-4;
